Tests fixtures
Move fixture files onto separated files with a special autoload for dev env.
