+++ test/Fixtures/FooInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Greg0ire\Enum\Tests\Fixtures;
+
+interface FooInterface
+{
+    const GOD = 'Dieu',
+        CHUCK = 'Chuck Norris',
+        GUITRY = 'Sacha Guitry';
+}
